@@ -108,33 +104,32 @@
     "redux": "^4.0.0",
     "redux-modules": "0.0.5",
     "redux-saga": "^0.16.0",
-    "remarkable": "2.0.1",
-    "sanitize-html": "^2.5.3",
     "sass-loader": "^13.2.0",
     "secure-random": "^1.1.1",
     "speakingurl": "^9.0.0",
     "string-to-stream": "^1.1.1",
     "style-loader": "^3.3.1",
-    "styled-components": "^5.3.8",
-    "webpack": "^5.75.0",
-    "webpack-dev-server": "^4.11.1",
+    "styled-components": "^6.1.17",
+    "webpack": "^5.99.7",
+    "webpack-dev-server": "^5.2.1",
     "whatwg-fetch": "^0.11.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.21.0",
-    "@babel/core": "^7.21.0",
-    "@babel/node": "^7.20.7",
-    "@babel/plugin-proposal-decorators": "^7.21.0",
-    "@babel/plugin-proposal-function-sent": "^7.18.6",
-    "@babel/plugin-proposal-throw-expressions": "^7.18.6",
-    "@babel/plugin-transform-runtime": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-react": "^7.18.6",
+    "@babel/cli": "^7.27.1",
+    "@babel/core": "^7.27.1",
+    "@babel/node": "^7.27.1",
+    "@babel/plugin-proposal-decorators": "^7.27.1",
+    "@babel/plugin-proposal-function-sent": "^7.27.1",
+    "@babel/plugin-proposal-throw-expressions": "^7.27.1",
+    "@babel/plugin-transform-runtime": "^7.27.1",
+    "@babel/preset-env": "^7.27.1",
+    "@babel/preset-react": "^7.27.1",
     "@red-mobile/cordova-plugin-shortcuts-android": "^1.0.1",
     "@storybook/addon-actions": "^6.5.14",
     "@storybook/addon-links": "^6.5.14",
     "@storybook/router": "^6.5.14",
-    "babel-plugin-styled-components": "^1.5.1",
+    "autoprefixer": "^10.4.21",
+    "babel-plugin-styled-components": "^2.1.4",
     "chai": "^4.1.2",
     "cheerio": "1.0.0-rc.10",
     "cordova-android": "^10.1.2",
