@@ -1,9 +1,10 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { IntlProvider } from 'react-intl'
-import { DEFAULT_LANGUAGE, LOCALE_COOKIE_KEY } from 'app/client_config';
 import tt from 'counterpart';
-import cookie from "react-cookie";
+import Cookies from 'universal-cookie'
+
+import { DEFAULT_LANGUAGE, LOCALE_COOKIE_KEY } from 'app/client_config'
 
 tt.registerTranslations('en', require('app/locales/en.json'));
 tt.registerTranslations('ru', require('app/locales/ru-RU.json'));
@@ -33,7 +34,8 @@ export default connect((state, props) => {
     let locale = state.user.get('locale')
 
     if (process.env.BROWSER) {
-        const l = cookie.load(LOCALE_COOKIE_KEY)
+        const cookies = new Cookies()
+        const l = cookies.get(LOCALE_COOKIE_KEY)
         if (l) locale = l;
     }
 
