@@ -1,10 +1,10 @@
-import koa_router from 'koa-router';
+import Router from 'koa-router'
 import React from 'react';
 import {routeRegex} from "app/ResolveRoute";
 import {api} from 'golos-lib-js';
 
 export default function useUserJson(app) {
-    const router = koa_router();
+    const router = new Router()
     app.use(router.routes());
 
     router.get(routeRegex.UserJson, async (ctx) => {
