@@ -49,7 +49,7 @@ module.exports = merge(baseConfig, {
                             postcssOptions: {
                                 plugins: [
                                     require('autoprefixer')({
-                                        browsers: ['> 1%', 'last 2 versions'],
+                                        overrideBrowserslist: ['> 1%', 'last 2 versions'],
                                     })
                                 ]
                             },
@@ -68,5 +68,8 @@ module.exports = merge(baseConfig, {
         compress: true,
         port: WEBPACK_PORT,
         hot: true,
+        client: {
+            overlay: false,
+        },
     },
 });
