@@ -36,7 +36,7 @@ module.exports = merge(baseConfig, {
                             postcssOptions: {
                                 plugins: [
                                     require('autoprefixer')({
-                                        browsers: ['> 1%', 'last 2 versions'],
+                                        overrideBrowserslist: ['> 1%', 'last 2 versions'],
                                     })
                                 ]
                             },
